Fix some codes

+++ src/Controle Servo Motor/servoMotor/controladores.h
@@ -2,8 +2,7 @@
 #define kd 0
 #define ki 0
 
-#include <math.h>
-double dErro_dt (int erroAtual)
+double dErro_dt (double erroAtual)
 {
   static          int   erroAnterior      = 0;
   static unsigned long  tempoAnterior     = 0;
@@ -17,7 +16,7 @@ double dErro_dt (int erroAtual)
   return dErro;
 }
 
-double iErro_dt (int erroAtual)
+double iErro_dt (double erroAtual)
 {
   static unsigned long  tempoAnterior     = 0;
          unsigned long  tempoAtual        = millis();
@@ -29,12 +28,12 @@ double iErro_dt (int erroAtual)
   return iErro;
 }
 
-int controlador_PID(int erro)
+double controlador_PID(double erro)
 {    
     return kp*erro + kd*dErro_dt(erro) + ki*iErro_dt(erro);
 }
 
-int controlador_SS(double referencia, double saidaPlanta, double ScA)//ScA tem que ser SATURADO !
+double controlador_SS(double referencia, double saidaPlanta, double ScA)//ScA tem que ser SATURADO !
 {
   static double EstadosA[2]   = {0, 0};  //Vetor de estados estimados
   static double integrador    =      0; //Valor do integrador
@@ -49,12 +48,12 @@ int controlador_SS(double referencia, double saidaPlanta, double ScA)//ScA tem q
   const  double Kw      =                               0.1                    ; //Ganho do Anti-WindUp
 
   double saidaEstimada  = C[0]*EstadosA[0]-C[1]*EstadosA[1];
-  EstadosA[0]           = A[0][0]*EstadosA[0]+A[0][1]*EstadosA[1]       +        B[0]*ScA        +        L[0]*(saidaPlanta-saidaEstimada); //Observador
-  EstadosA[1]           = A[1][0]*EstadosA[0]+A[1][1]*EstadosA[1]       +        B[1]*ScA        +        L[1]*(saidaPlanta-saidaEstimada); //Observador
+  EstadosA[0]           = A[0][0]*EstadosA[0] + A[0][1]*EstadosA[1]       +        B[0]*ScA        +        L[0]*(saidaPlanta-saidaEstimada); //Observador
+  EstadosA[1]           = A[1][0]*EstadosA[0] + A[1][1]*EstadosA[1]       +        B[1]*ScA        +        L[1]*(saidaPlanta-saidaEstimada); //Observador
          
-  integrador += (referencia - saidaPlanta)*taxAmo + anti_windup*Kw;      //Integra o erro em função do tempo e soma ação anti-windup
+  integrador += (referencia - saidaPlanta)*taxAmo + anti_windup*Kw;              //Integra o erro em função do tempo e soma ação anti-windup
   
-  double Sc = integrador*Ka - K[0]*EstadosA[0] -  K[1]*EstadosA[1];      //Controlador retorna Sc
+  double Sc = integrador*Ka - K[0]*EstadosA[0] -  K[1]*EstadosA[1];              //Controlador retorna Sc
   anti_windup = ScA - Sc;
-  return (int)Sc;
+  return Sc;
 }
